+++ src/main/java/pl/coderslab/finalproject/child/ChildController.java
@@ -8,7 +8,9 @@ import pl.coderslab.finalproject.announcement.Announcement;
 import pl.coderslab.finalproject.group.GroupDao;
 import pl.coderslab.finalproject.parent.Parent;
 import pl.coderslab.finalproject.parent.ParentDao;
+import pl.coderslab.finalproject.user.Role;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 
@@ -19,7 +21,6 @@ public class ChildController {
     private final ParentDao parentDao;
     private final GroupDao groupDao;
 
-
     public ChildController(ChildDao childDao, ParentDao parentDao, GroupDao groupDao) {
         this.childDao = childDao;
         this.parentDao = parentDao;
@@ -27,12 +28,16 @@ public class ChildController {
     }
 
     @GetMapping("/add")
-    public String addChild(Model model) {
-        Child child = new Child();
-        model.addAttribute("child", child);
-        model.addAttribute("groups", groupDao.findAll());
-        model.addAttribute("parents", parentDao.findAll());
-        return "child/form";
+    public String addChild(Model model, HttpSession session) {
+        if (session.getAttribute("role") != null && session.getAttribute("role").equals(Role.ADMIN)) {
+            Child child = new Child();
+            model.addAttribute("child", child);
+            model.addAttribute("groups", groupDao.findAll());
+            model.addAttribute("parents", parentDao.findAll());
+            return "child/form";
+        } else {
+            return "user/noAccess";
+        }
     }
 
     @PostMapping("/add")
@@ -40,6 +45,7 @@ public class ChildController {
         childDao.saveChild(child);
         return "redirect:/child/list";
     }
+
     @PostMapping("/delete")
     public String deleteChild(Child child) {
         childDao.deleteChild(child);
@@ -47,34 +53,51 @@ public class ChildController {
     }
 
 
-   @GetMapping("/list")
-    public String getList(Model model) {
-        List<Child> children = childDao.findAll();
-        model.addAttribute("children", children);
-        return "child/list";
+    @GetMapping("/list")
+    public String getList(Model model, HttpSession session) {
+        if (session.getAttribute("role") != null && session.getAttribute("role").equals(Role.ADMIN)) {
+            List<Child> children = childDao.findAll();
+            model.addAttribute("children", children);
+            return "child/list";
+        } else {
+            return "user/noAccess";
+        }
     }
 
-@RequestMapping("delete/{id}")
-public String deleteChild(@PathVariable Long id) {
-    Child child = childDao.findById(id);
-    childDao.delete(id);
-    return "redirect:/child/list";
-}
-
-   @GetMapping("/update/{id}")
-    public String updateChild(@PathVariable Long id, Model model) {
-        Child child = childDao.findById(id);
-        model.addAttribute("child", child);
-        model.addAttribute("groups", groupDao.findAll());
-        model.addAttribute("parents", parentDao.findAll());
-        return "child/editForm";
+    @RequestMapping("delete/{id}")
+    public String deleteChild(@PathVariable Long id, HttpSession session) {
+        if (session.getAttribute("role") != null && session.getAttribute("role").equals(Role.ADMIN)) {
+            Child child = childDao.findById(id);
+            childDao.delete(id);
+            return "redirect:/child/list";
+        } else {
+            return "user/noAccess";
+        }
     }
-    @PostMapping ("/update")
-    public String updatedChild(@ModelAttribute Child child, Model model){
-        childDao.saveChild(child);
-        List <Child> children = childDao.findAll();
-        model.addAttribute("children", children);
-        return "child/list";
+
+    @GetMapping("/update/{id}")
+    public String updateChild(@PathVariable Long id, Model model,HttpSession session) {
+        if (session.getAttribute("role") != null && session.getAttribute("role").equals(Role.ADMIN)) {
+            Child child = childDao.findById(id);
+            model.addAttribute("child", child);
+            model.addAttribute("groups", groupDao.findAll());
+            model.addAttribute("parents", parentDao.findAll());
+            return "child/editForm";
+        } else {
+            return "user/noAccess";
+        }
+    }
+
+    @PostMapping("/update")
+    public String updatedChild(@ModelAttribute Child child, Model model, HttpSession session) {
+        if (session.getAttribute("role") != null && session.getAttribute("role").equals(Role.ADMIN)) {
+            childDao.saveChild(child);
+            List<Child> children = childDao.findAll();
+            model.addAttribute("children", children);
+            return "child/list";
+        } else {
+            return "user/noAccess";
+        }
     }
 }
